Float/decimal precision corrected

+++ homeloop/db/migrate/20170608194905_create_queries.rb
@@ -2,8 +2,8 @@ class CreateQueries < ActiveRecord::Migration[5.1]
   def change
     create_table :queries do |t|
       t.string :email
-      t.float :lat
-      t.float :lon
+      t.decimal :lat
+      t.decimal :lon
       t.integer :res_count
 
       t.timestamps
